feat: Added basic classes for SFU and Producer, added websocket server connection, added events for basic RTCPeerConnection
Added Producer class that creates a new RTCPeerConnection, adds callbacks for RTC handshake. Added SFU class that keeps a collection of producers and a socket, and binds events to the socket. Also added a Consumer class that will be used by clients to get streams.

Consider suggestion to refactor to have a Client class that holds a single Producer and a collection of consumers. SFU would then have a collection of clients.

Co-authored-by: Tyler Frye <tyler.frye42@gmail.com>
Co-authored-by: Jay Jayabose <jayabose@gmail.com>